requirements.txt
Django==4.2.1
django-rest-framework-social-oauth2==1.1.0
psycopg2-binary==2.9.6
requests==2.31.0